backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
pydantic==2.5.2
gspread==6.2.0
oauth2client==4.1.3
faiss-cpu==1.7.4 